+++ rmr/settings.py
@@ -27,6 +27,8 @@ DEBUG = False
 
 ALLOWED_HOSTS = []
 
+CACHE_MIDDLEWARE_SECONDS = 0
+
 
 # Application definition
 
@@ -35,6 +37,8 @@ INSTALLED_APPS = (
 )
 
 MIDDLEWARE_CLASSES = (
+    'rmr.extensions.middleware.cache.UpdateCacheMiddleware',
+    'django.middleware.http.ConditionalGetMiddleware',
     'rmr.extensions.middleware.json.RequestDecoder',
 )
 
@@ -64,3 +68,27 @@ USE_I18N = True
 USE_L10N = True
 
 USE_TZ = True
+
+
+# Logging
+# https://docs.djangoproject.com/en/1.8/ref/settings/#id11
+
+LOGGING = {
+    'version': 1,
+    'disable_existing_loggers': False,
+    'handlers': {
+        'null': {
+            'class': 'logging.NullHandler',
+        },
+    },
+    'loggers': {
+        'django': {
+            'handlers': ['null'],
+            'level': 'ERROR',
+        },
+        'rmr': {
+            'handlers': ['null'],
+            'level': 'ERROR',
+        },
+    },
+}